+++ src/main/resources/mapper/PostMapper.xml
@@ -475,4 +475,7 @@
         )
         WHERE post_id = #{postId}
     </update>
+
+
+
 </mapper>
