+++ src/main/webapp/WEB-INF/views/mypage/myBoard.jsp
@@ -1,0 +1,17 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: kosa
+  Date: 2025-09-10
+  Time: 오후 8:20
+  To change this template use File | Settings | File Templates.
+--%>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<html>
+<head>
+    <title>Titleasdasd</title>
+</head>
+  <h1>내가 ㅡsadsa</h1>
+<body>
+
+</body>
+</html>
