+++ src/main/webapp/WEB-INF/views/post/content.jsp
@@ -1,6 +1,6 @@
 <%@ page contentType="text/html; charset=UTF-8" %>
 
-<h2>게시판 페이지</h2>
+<h2>게시판 페이지asdasdas</h2>
 <div class="grid-container">
     게시판 페이지 내용입니다.
 </div>
